moved loader logic into extra module in fun/loader
